:truck: Rename Page to Register and move it to pages folder
